Modified several mpi4py routines and included new mpi4py features
"For the mpisend, recv, etc, I would check the input type and use the appropriate send or recv, instead of relying on the lmpi4pynp flag. The user wouldn't have to worry about it then, and the code would use the faster option when possible. Also, for mpirecv, I would have put the data argument last, and called it "out" to be consistent with what numpy does. If not supplied, it could be created in the routine as needed if the faster routine is being used."

Using the "fast" Send and Recv routines directly, only by checking if the object is a numpy array, did not work properly and warp got buggy. Therefor i rewrote the complete implementation and changed the way send and recv is used for the mpi4py case. I now check if the object is a numpy array. If it is a numpy array, it is sent by using the "fast" routines, but also the shape and the dtype of the array is sent by using the "slow" routines. The receiving process then first receives the shape and the dtype. Then an empty array is created based on this information. This empty array is then used as the "data" ("out") argument to receive the array via the "fast" routine Recv. I don't know if this is a good way to solve the problem, but it is fast and there is no need for the input arguments "lmpi4pynp" or "out"/"data" anymore.

"There are all the various places with the replacements like mpisend for comm_world.send. In all these places, the explicit comm_world is not needed (since the one imported from warp is always used) and might as well be removed for simplicity."
Removed all comm = comm_world.

"A function called mpicommcreate should be added to parallel.py and then used in quasistatic.py, to hide the try/except block when creating mympigroup."
A function mpicommcreate(group) is now available in parallel.py

"For consistency, a routine mpiisend should be created instead of calling comm_world.isend."
mpiisend() is now available in parallel.py

"runcounter.py is calling comm_world.barrier(). This should be changed to barrier()."
Done.

"fieldsolver.py is calling comm_world.scatter. Is there an equivalent in mpi4py? A routine mpiscatter should be added to parallel.py to replace the direct call."
mpiscatter() is now available in parallel.py

"For synchronizeQueuedOutput, while I know all of the options are not available in mpi4py, I would have preferred keeping the same interface, though this is minor. Why is exec being used, instead of importing sys?"
I found this code snippet somewhere in the internet and just added it. You can change it, as I don't really know how to do this in a good way.

Performance test of pyMPI/mpi4py:

send and recv of an array random.rand(100, 100, 100) from one process to another:

mpi4py slow routines:
0.017

mpi4py hybrid fast routines (current implementation):
0.0020

mpi4py fastest routines (predefining the array on both processes):
0.0016

pyMPI:
0.0036