Fix MPI error with the laser antenna
When running an MPI simulation and when the laser antenna
is exactly at the border between to processors, there is a
bug which results in an twice higher a0 than expected.

This is because the if condition (whether or not the antenna is
inside the domain for the present processor) is met for both the
neighboring processors! The reason for this absurd behavior is that
the quantity zmax of the left processor is not equal to zmin of the
right processor (there are small, machine-precision differences
which come from the fact that zmin and zmax are not initialized in
the same way).

The present commit changes the initialization of zmin and zmax
so that the zmin of the right processor is always **exactly** equal
to the zmax of the left processor. This fixes the above mention bug.