Modified boosted_particle_diag.
The approach here is to have each core to process as much as it can, then gather at the end before writing to the disk. This helps in reducing memory requirement in creating another array of the quantities of the previous time step and also time requirement to transfer this array to the processor 0.